requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
billiard==4.2.0
bip32==3.4
bip39==0.0.2
bitarray==2.9.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coincurve==18.0.0
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cytoolz==0.12.3
Django==5.0.4
cryptography
django-admin-honeypot-updated-2021==1.2.0
django-allauth==0.62.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-countries==7.6.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-htmx==1.17.3
django-ipware==7.0.1
django-js-asset==2.2.0
django-social-share==2.3.0
django-taggit==5.0.1
django-timezone-field==6.1.0
easy-thumbnails==2.8.5
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.1
eth-utils==4.1.0
eth_abi==5.1.0
Faker==24.14.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
lru-dict==1.2.0
mnemonic==0.21
multidict==6.0.5
parsimonious==0.10.0
pillow==10.3.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-ipware==3.0.0
pyuca==1.2
pyunormalize==15.1.0
qrcode==7.4.2
redis==5.0.4
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.0
six==1.16.0
sqlparse==0.5.0
sweetify==2.3.1
toolz==0.12.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
web3==6.17.2
websockets==12.0
yarl==1.9.4
